refactor: Split util functions
